pyproject.toml
[tool.poetry]
name = "job-aggregator"
version = "0.1.0"
description = ""
authors = ["BogdanovaMF <122004102+BogdanovaMF@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "job_aggregator"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "1.5.3"
parsel = "1.7.0"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.7"
mypy = "^1.2.0"
telethon = "^1.28.5"
numpy = "^1.24.3"

[tool.poetry.group.dev.dependencies]
requests = "2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
